Update V005__CREATE TABLE user.sql

+++ V005__CREATE TABLE user.sql
@@ -1,5 +1,5 @@
 CREATE TABLE user(
-    id BIGSERIAL NOT NULL,
+    id integer NOT NULL,
     first_name VARCHAR(250) NOT NULL,
     last_name VARCHAR(250) NOT NULL,
     email VARCHAR(250) NOT NULL,
